refaktorere etteroppgjør pakker
