+++ apps/etterlatte-behandling/src/main/kotlin/behandling/etteroppgjoer/inntektskomponent/InntektskomponentModel.kt
@@ -1,0 +1,51 @@
+package no.nav.etterlatte.behandling.etteroppgjoer
+
+import java.math.BigDecimal
+import java.time.YearMonth
+
+data class AInntekt(
+    val aar: Int,
+    val inntektsmaaneder: List<AInntektMaaned>,
+) {
+    init {
+        require(inntektsmaaneder.all { it.maaned.year == aar }) {
+            "Alle inntektsmaaneder må ha inntektsår = $aar, men fant: ${inntektsmaaneder.map { it.maaned.year }}"
+        }
+    }
+
+    companion object {
+        fun stub(
+            aar: Int = 2024,
+            aarsinntekt: Int = 300000,
+        ) = AInntekt(
+            aar = 2024,
+            // inntektsmaaneder = listOf("01", "02", "03", "04", "05", "06", "07", "08", "09", "10", "11", "12").map {
+            inntektsmaaneder =
+                listOf(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12).map {
+                    AInntektMaaned(
+                        // maaned = "$aar-$it",
+                        maaned = YearMonth.of(aar, it),
+                        inntekter =
+                            listOf(
+                                Inntekt(
+                                    beloep = BigDecimal(aarsinntekt / 12),
+                                    beskrivelse = "En månedsinntekt",
+                                ),
+                            ),
+                        summertBeloep = BigDecimal(aarsinntekt / 12),
+                    )
+                },
+        )
+    }
+}
+
+data class AInntektMaaned(
+    val maaned: YearMonth,
+    val inntekter: List<Inntekt>,
+    val summertBeloep: BigDecimal,
+)
+
+data class Inntekt(
+    val beloep: BigDecimal,
+    val beskrivelse: String,
+)
