+++ apps/etterlatte-behandling/src/main/kotlin/brev/BrevService.kt
@@ -3,8 +3,8 @@ package no.nav.etterlatte.brev
 import io.ktor.client.call.body
 import io.ktor.client.plugins.ResponseException
 import no.nav.etterlatte.behandling.BehandlingService
-import no.nav.etterlatte.behandling.etteroppgjoer.forbehandling.EtteroppgjoerForbehandlingBrevService
-import no.nav.etterlatte.behandling.etteroppgjoer.revurdering.EtteroppgjoerRevurderingBrevService
+import no.nav.etterlatte.behandling.etteroppgjoer.brev.EtteroppgjoerForbehandlingBrevService
+import no.nav.etterlatte.behandling.etteroppgjoer.brev.EtteroppgjoerRevurderingBrevService
 import no.nav.etterlatte.behandling.klienter.BrevApiKlient
 import no.nav.etterlatte.behandling.klienter.VedtakKlient
 import no.nav.etterlatte.behandling.vedtaksbehandling.BehandlingMedBrevService
